test(css-playground): split Next.js consumer into isolated per-aspect workspaces
Replace the single @css-playground/next app with three isolated Next workspaces
so cross-route bundling optimizations cannot bleed between scenarios:
- next-server-component: default Server Component page with a direct import
- next-client-component: 'use client' page with a direct import
- next-dynamic: Server Component page using next/dynamic

Each keeps the createConfig() import in next.config.ts (the #12825 Node-eval
repro) and builds with next build --turbopack.
