+++ css-playground/next-client-component/app/layout.tsx
@@ -1,0 +1,13 @@
+import type {ReactNode} from 'react'
+
+export const metadata = {
+  title: '@css-playground/next-client-component',
+}
+
+export default function RootLayout({children}: {children: ReactNode}) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  )
+}
