+++ css-playground/next-dynamic/app/layout.tsx
@@ -1,7 +1,7 @@
 import type {ReactNode} from 'react'
 
 export const metadata = {
-  title: '@css-playground/next',
+  title: '@css-playground/next-dynamic',
 }
 
 export default function RootLayout({children}: {children: ReactNode}) {
