+++ css-playground/next-dynamic/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@css-playground/next",
+  "name": "@css-playground/next-dynamic",
   "version": "1.0.0",
   "private": true,
   "scripts": {
