+++ css-playground/next-server-component/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@css-playground/next-server-component",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "build": "next build --turbopack",
+    "clean": "rimraf .next",
+    "test": "next build --turbopack"
+  },
+  "devDependencies": {
+    "@types/node": "catalog:",
+    "@types/react": "^19.2.17",
+    "@types/react-dom": "^19.2.3",
+    "next": "^16.2.9",
+    "react": "^19.2.7",
+    "react-dom": "^19.2.7",
+    "rimraf": "^6.1.3",
+    "sanity-css-vanilla-extract-test": "workspace:*"
+  }
+}
